implemented grpc client

+++ src/main/java/com/example/quickstart/repository/WalletRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface WalletRepository extends JpaRepository<WalletModel,Long> {
+public interface WalletRepository extends JpaRepository<WalletModel,Integer> {
 
 }
